requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
Flask==3.0.3
Flask-Cors==5.0.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-dialogflow==2.33.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.0.2
pip-autoremove==0.10.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scipy==1.13.1
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
